requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==4.0.3
django-filter==21.1
django-heroku==0.3.1
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
Pillow==9.1.0
psycopg2==2.9.3
python-dateutil==1.5
requests==2.27.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.9
users==1.0.dev0
whitenoise==6.0.0
